requirements/dev.txt
-r tests.txt

django-debug-toolbar==1.3.2
  sqlparse==0.1.15
